+++ config/default.ts
@@ -1,5 +1,8 @@
 export default {
-    port: 3000,
-    dbURL: 'mongodb+srv://Frodo:kjhlcfehjy@cluster0.dqg2t1i.mongodb.net/',
-    salt: 'kjhlcfehjy'
-}
+  port: 3000,
+  dbURL: "mongodb+srv://Frodo:kjhlcfehjy@cluster0.dqg2t1i.mongodb.net/",
+  salt: "kjhlcfehjy",
+  cloud_name: "do07m8yye",
+  api_key: "661317545416412",
+  api_secret: "zCVa6yQB5UbJNK2_LMlxwKGrKJc"
+};
