+++ src/interfaces/IRequestOwner.ts
@@ -1,0 +1,4 @@
+export interface IRequestOwner {
+  userId: string;
+  email: string;
+}
