+++ src/middlewares/fileUploadMiddleware.ts
@@ -1,12 +1,17 @@
 import { Request } from "express";
 import multer from "multer";
 const cloudinary = require("cloudinary").v2;
+import config from "config";
 const { CloudinaryStorage } = require("multer-storage-cloudinary");
 
+const cloud_name = config.get<string>("cloud_name");
+const api_key = config.get<string>("api_key");
+const api_secret = config.get<string>("api_secret");
+
 cloudinary.config({
-  cloud_name: "do07m8yye",
-  api_key: "661317545416412",
-  api_secret: "zCVa6yQB5UbJNK2_LMlxwKGrKJc",
+  cloud_name: cloud_name,
+  api_key: api_key,
+  api_secret: api_secret,
 });
 
 const storage = new CloudinaryStorage({
